TINY-4847: Move types to Ui module
